refactor: extract cpptrace related symbols into mimic++_ext/stacktrace/cpptrace.hpp
